+++ CHANGELOG.md
@@ -10,10 +10,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 - `CmpEqBy`, `CmpEq`, `GTBy`, `GT`, `LTBy`, `LT`, `GTEBy`, `GTE`, `LTEBy`, and `LTE` inequality checks
 - `MinBy`, `MaxBy`, `Min`, and `Max` semigroups
 - `Product2-8` interfaces, representing general product types
-- `Union`, a semigroup that behaves like a lazy set union on `Iterable`s
+- `Union`, a monoid that behaves like a lazy set union on `Iterable`s
 - `Difference`, a semigroup that behaves like a partially lazy set difference on `Iterable`s
 - `LambdaMap`, extension point for `j.u.Map`, similar to `LambdaIterable`
 - `Sequence#sequence` overloads for `j.u.Map` that traverse via intermediate `LambdaMap` instances
+- `Intersection`, a semigroup that behaves like a lazy set intersection on `Iterable`s
 
 ### Changed
 - `Tuple2-8` now implement `Product2-8`